# ad controller: lift nodeAttachedTo.mountedByNode